requirements.txt
openai==1.33.0
neo4j-driver==5.21.0
streamlit==1.35.0